add _int suffix to all the internal files
