+++ src/canonmn_int.cpp
