+++ src/casiomn_int.cpp
