+++ src/fujimn_int.cpp
