+++ src/makernote_int.cpp
