+++ src/minoltamn_int.cpp
