+++ src/nikonmn_int.cpp
