+++ src/panasonicmn_int.cpp
