+++ src/pngchunk_int.cpp
