+++ src/samsungmn_int.cpp
