+++ src/sigmamn_int.cpp
