+++ src/tags_int.cpp
@@ -1,212 +1,6 @@
-// ***************************************************************** -*- C++ -*-
-/*
- * Copyright (C) 2004-2017 Andreas Huggel <ahuggel@gmx.net>
- *
- * This program is part of the Exiv2 distribution.
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, 5th Floor, Boston, MA 02110-1301 USA.
- */
-/*
-  File:      tags.cpp
-  Author(s): Andreas Huggel (ahu) <ahuggel@gmx.net>
-             Gilles Caulier (gc) <caulier dot gilles at gmail dot com>
-  History:   15-Jan-04, ahu: created
-             21-Jan-05, ahu: added MakerNote TagInfo registry and related code
- */
-// *****************************************************************************
-// included header files
-#include "types.hpp"
-#include "tags.hpp"
 #include "tags_int.hpp"
-#include "error.hpp"
-#include "futils.hpp"
-#include "value.hpp"
-#include "convert.hpp"
+
 #include "i18n.h"                // NLS support.
-
-#include "canonmn_int.hpp"
-#include "casiomn_int.hpp"
-#include "fujimn_int.hpp"
-#include "minoltamn_int.hpp"
-#include "nikonmn_int.hpp"
-#include "olympusmn_int.hpp"
-#include "panasonicmn_int.hpp"
-#include "pentaxmn_int.hpp"
-#include "samsungmn_int.hpp"
-#include "sigmamn_int.hpp"
-#include "sonymn_int.hpp"
-
-#include <iostream>
-#include <iomanip>
-#include <sstream>
-#include <utility>
-#include <cstdlib>
-#include <cassert>
-#include <cmath>
-#include <cstring>
-
-// *****************************************************************************
-// local declarations
-namespace {
-    // Print version string from an intermediate string
-    std::ostream& printVersion(std::ostream& os, const std::string& str);
-}
-
-// *****************************************************************************
-// class member definitions
-namespace Exiv2 {
-
-    using namespace Internal;
-
-    //! List of all known Exif groups. Important: Group name (3rd column) must be unique!
-    extern const GroupInfo groupInfo[] = {
-        { ifdIdNotSet,     "Unknown IFD", "Unknown", 0 },
-        { ifd0Id,          "IFD0",      "Image",        ifdTagList                     },
-        { ifd1Id,          "IFD1",      "Thumbnail",    ifdTagList                     },
-        { ifd2Id,          "IFD2",      "Image2",       ifdTagList                     },
-        { ifd3Id,          "IFD3",      "Image3",       ifdTagList                     },
-        { exifId,          "Exif",      "Photo",        exifTagList                    },
-        { gpsId,           "GPSInfo",   "GPSInfo",      gpsTagList                     },
-        { iopId,           "Iop",       "Iop",          iopTagList                     },
-        { mpfId,           "MPF",       "MpfInfo",      mpfTagList                     },
-        { subImage1Id,     "SubImage1", "SubImage1",    ifdTagList                     },
-        { subImage2Id,     "SubImage2", "SubImage2",    ifdTagList                     },
-        { subImage3Id,     "SubImage3", "SubImage3",    ifdTagList                     },
-        { subImage4Id,     "SubImage4", "SubImage4",    ifdTagList                     },
-        { subImage5Id,     "SubImage5", "SubImage5",    ifdTagList                     },
-        { subImage6Id,     "SubImage6", "SubImage6",    ifdTagList                     },
-        { subImage7Id,     "SubImage7", "SubImage7",    ifdTagList                     },
-        { subImage8Id,     "SubImage8", "SubImage8",    ifdTagList                     },
-        { subImage9Id,     "SubImage9", "SubImage9",    ifdTagList                     },
-        { subThumb1Id,     "SubThumb1", "SubThumb1",    ifdTagList                     },
-        { panaRawId,       "PanaRaw",   "PanasonicRaw", PanasonicMakerNote::tagListRaw },
-        { mnId,            "Makernote", "MakerNote",    mnTagList                      },
-        { canonId,         "Makernote", "Canon",        CanonMakerNote::tagList        },
-        { canonCsId,       "Makernote", "CanonCs",      CanonMakerNote::tagListCs      },
-        { canonSiId,       "Makernote", "CanonSi",      CanonMakerNote::tagListSi      },
-        { canonCfId,       "Makernote", "CanonCf",      CanonMakerNote::tagListCf      },
-        { canonPiId,       "Makernote", "CanonPi",      CanonMakerNote::tagListPi      },
-        { canonTiId,       "Makernote", "CanonTi",      CanonMakerNote::tagListTi      },
-        { canonFiId,       "Makernote", "CanonFi",      CanonMakerNote::tagListFi      },
-        { canonPaId,       "Makernote", "CanonPa",      CanonMakerNote::tagListPa      },
-        { canonPrId,       "Makernote", "CanonPr",      CanonMakerNote::tagListPr      },
-        { casioId,         "Makernote", "Casio",        CasioMakerNote::tagList        },
-        { casio2Id,        "Makernote", "Casio2",       Casio2MakerNote::tagList       },
-        { fujiId,          "Makernote", "Fujifilm",     FujiMakerNote::tagList         },
-        { minoltaId,       "Makernote", "Minolta",      MinoltaMakerNote::tagList      },
-        { minoltaCs5DId,   "Makernote", "MinoltaCs5D",  MinoltaMakerNote::tagListCs5D  },
-        { minoltaCs7DId,   "Makernote", "MinoltaCs7D",  MinoltaMakerNote::tagListCs7D  },
-        { minoltaCsOldId,  "Makernote", "MinoltaCsOld", MinoltaMakerNote::tagListCsStd },
-        { minoltaCsNewId,  "Makernote", "MinoltaCsNew", MinoltaMakerNote::tagListCsStd },
-        { nikon1Id,        "Makernote", "Nikon1",       Nikon1MakerNote::tagList       },
-        { nikon2Id,        "Makernote", "Nikon2",       Nikon2MakerNote::tagList       },
-        { nikon3Id,        "Makernote", "Nikon3",       Nikon3MakerNote::tagList       },
-        { nikonPvId,       "Makernote", "NikonPreview", ifdTagList                     },
-        { nikonVrId,       "Makernote", "NikonVr",      Nikon3MakerNote::tagListVr     },
-        { nikonPcId,       "Makernote", "NikonPc",      Nikon3MakerNote::tagListPc     },
-        { nikonWtId,       "Makernote", "NikonWt",      Nikon3MakerNote::tagListWt     },
-        { nikonIiId,       "Makernote", "NikonIi",      Nikon3MakerNote::tagListIi     },
-        { nikonAfId,       "Makernote", "NikonAf",      Nikon3MakerNote::tagListAf     },
-        { nikonAf2Id,      "Makernote", "NikonAf2",     Nikon3MakerNote::tagListAf2    },
-        { nikonAFTId,      "Makernote", "NikonAFT",     Nikon3MakerNote::tagListAFT    },
-        { nikonFiId,       "Makernote", "NikonFi",      Nikon3MakerNote::tagListFi     },
-        { nikonMeId,       "Makernote", "NikonMe",      Nikon3MakerNote::tagListMe     },
-        { nikonFl1Id,      "Makernote", "NikonFl1",     Nikon3MakerNote::tagListFl1    },
-        { nikonFl2Id,      "Makernote", "NikonFl2",     Nikon3MakerNote::tagListFl2    },
-        { nikonFl3Id,      "Makernote", "NikonFl3",     Nikon3MakerNote::tagListFl3    },
-        { nikonSi1Id,      "Makernote", "NikonSiD80",   Nikon3MakerNote::tagListSi1    },
-        { nikonSi2Id,      "Makernote", "NikonSiD40",   Nikon3MakerNote::tagListSi2    },
-        { nikonSi3Id,      "Makernote", "NikonSiD300a", Nikon3MakerNote::tagListSi3    },
-        { nikonSi4Id,      "Makernote", "NikonSiD300b", Nikon3MakerNote::tagListSi4    },
-        { nikonSi5Id,      "Makernote", "NikonSi02xx",  Nikon3MakerNote::tagListSi5    },
-        { nikonSi6Id,      "Makernote", "NikonSi01xx",  Nikon3MakerNote::tagListSi5    },
-        { nikonCb1Id,      "Makernote", "NikonCb1",     Nikon3MakerNote::tagListCb1    },
-        { nikonCb2Id,      "Makernote", "NikonCb2",     Nikon3MakerNote::tagListCb2    },
-        { nikonCb2aId,     "Makernote", "NikonCb2a",    Nikon3MakerNote::tagListCb2a   },
-        { nikonCb2bId,     "Makernote", "NikonCb2b",    Nikon3MakerNote::tagListCb2b   },
-        { nikonCb3Id,      "Makernote", "NikonCb3",     Nikon3MakerNote::tagListCb3    },
-        { nikonCb4Id,      "Makernote", "NikonCb4",     Nikon3MakerNote::tagListCb4    },
-        { nikonLd1Id,      "Makernote", "NikonLd1",     Nikon3MakerNote::tagListLd1    },
-        { nikonLd2Id,      "Makernote", "NikonLd2",     Nikon3MakerNote::tagListLd2    },
-        { nikonLd3Id,      "Makernote", "NikonLd3",     Nikon3MakerNote::tagListLd3    },
-        { olympusId,       "Makernote", "Olympus",      OlympusMakerNote::tagList      },
-        { olympus2Id,      "Makernote", "Olympus2",     OlympusMakerNote::tagList      },
-        { olympusCsId,     "Makernote", "OlympusCs",    OlympusMakerNote::tagListCs    },
-        { olympusEqId,     "Makernote", "OlympusEq",    OlympusMakerNote::tagListEq    },
-        { olympusRdId,     "Makernote", "OlympusRd",    OlympusMakerNote::tagListRd    },
-        { olympusRd2Id,    "Makernote", "OlympusRd2",   OlympusMakerNote::tagListRd2   },
-        { olympusIpId,     "Makernote", "OlympusIp",    OlympusMakerNote::tagListIp    },
-        { olympusFiId,     "Makernote", "OlympusFi",    OlympusMakerNote::tagListFi    },
-        { olympusFe1Id,    "Makernote", "OlympusFe1",   OlympusMakerNote::tagListFe    },
-        { olympusFe2Id,    "Makernote", "OlympusFe2",   OlympusMakerNote::tagListFe    },
-        { olympusFe3Id,    "Makernote", "OlympusFe3",   OlympusMakerNote::tagListFe    },
-        { olympusFe4Id,    "Makernote", "OlympusFe4",   OlympusMakerNote::tagListFe    },
-        { olympusFe5Id,    "Makernote", "OlympusFe5",   OlympusMakerNote::tagListFe    },
-        { olympusFe6Id,    "Makernote", "OlympusFe6",   OlympusMakerNote::tagListFe    },
-        { olympusFe7Id,    "Makernote", "OlympusFe7",   OlympusMakerNote::tagListFe    },
-        { olympusFe8Id,    "Makernote", "OlympusFe8",   OlympusMakerNote::tagListFe    },
-        { olympusFe9Id,    "Makernote", "OlympusFe9",   OlympusMakerNote::tagListFe    },
-        { olympusRiId,     "Makernote", "OlympusRi",    OlympusMakerNote::tagListRi    },
-        { panasonicId,     "Makernote", "Panasonic",    PanasonicMakerNote::tagList    },
-        { pentaxDngId,     "Makernote", "PentaxDng",    PentaxMakerNote::tagList       },
-        { pentaxId,        "Makernote", "Pentax",       PentaxMakerNote::tagList       },
-        { samsung2Id,      "Makernote", "Samsung2",     Samsung2MakerNote::tagList     },
-        { samsungPvId,     "Makernote", "SamsungPreview", ifdTagList                   },
-        { samsungPwId,     "Makernote", "SamsungPictureWizard", Samsung2MakerNote::tagListPw },
-        { sigmaId,         "Makernote", "Sigma",        SigmaMakerNote::tagList        },
-        { sony1Id,         "Makernote", "Sony1",        SonyMakerNote::tagList         },
-        { sony2Id,         "Makernote", "Sony2",        SonyMakerNote::tagList         },
-        { sonyMltId,       "Makernote", "SonyMinolta",  MinoltaMakerNote::tagList      },
-        { sony1CsId,       "Makernote", "Sony1Cs",      SonyMakerNote::tagListCs       },
-        { sony1Cs2Id,      "Makernote", "Sony1Cs2",     SonyMakerNote::tagListCs2      },
-        { sony1MltCs7DId,  "Makernote", "Sony1MltCs7D", MinoltaMakerNote::tagListCs7D  },
-        { sony1MltCsOldId, "Makernote", "Sony1MltCsOld",MinoltaMakerNote::tagListCsStd },
-        { sony1MltCsNewId, "Makernote", "Sony1MltCsNew",MinoltaMakerNote::tagListCsStd },
-        { sony1MltCsA100Id,"Makernote","Sony1MltCsA100",MinoltaMakerNote::tagListCsA100},
-        { sony2CsId,       "Makernote", "Sony2Cs",      SonyMakerNote::tagListCs       },
-        { sony2Cs2Id,      "Makernote", "Sony2Cs2",     SonyMakerNote::tagListCs2      },
-        { lastId,          "(Last IFD info)", "(Last IFD item)", 0 }
-    };
-
-    //! List of all defined Exif sections.
-    extern const SectionInfo sectionInfo[] = {
-        { sectionIdNotSet, "(UnknownSection)",     N_("Unknown section")              },
-        { imgStruct,       "ImageStructure",       N_("Image data structure")         },
-        { recOffset,       "RecordingOffset",      N_("Recording offset")             },
-        { imgCharacter,    "ImageCharacteristics", N_("Image data characteristics")   },
-        { otherTags,       "OtherTags",            N_("Other data")                   },
-        { exifFormat,      "ExifFormat",           N_("Exif data structure")          },
-        { exifVersion,     "ExifVersion",          N_("Exif version")                 },
-        { imgConfig,       "ImageConfig",          N_("Image configuration")          },
-        { userInfo,        "UserInfo",             N_("User information")             },
-        { relatedFile,     "RelatedFile",          N_("Related file")                 },
-        { dateTime,        "DateTime",             N_("Date and time")                },
-        { captureCond,     "CaptureConditions",    N_("Picture taking conditions")    },
-        { gpsTags,         "GPS",                  N_("GPS information")              },
-        { iopTags,         "Interoperability",     N_("Interoperability information") },
-        { mpfTags,         "MPF",                  N_("CIPA Multi-Picture Format")    },
-        { makerTags,       "Makernote",            N_("Vendor specific information")  },
-        { dngTags,         "DngTags",              N_("Adobe DNG tags")               },
-        { panaRaw,         "PanasonicRaw",         N_("Panasonic RAW tags")           },
-        { tiffEp,          "TIFF/EP",              N_("TIFF/EP tags")                 },
-        { tiffPm6,         "TIFF&PM6",             N_("TIFF PageMaker 6.0 tags")      },
-        { adobeOpi,        "AdobeOPI",             N_("Adobe OPI tags")               },
-        { lastSectionId,   "(LastSection)",        N_("Last section")                 }
-    };
-
-} // namespace Exiv2
 
 namespace Exiv2 {
     namespace Internal {
@@ -2253,11 +2047,6 @@ namespace Exiv2 {
         return mnTagInfo;
     }
 
-    // Unknown Tag
-    static const TagInfo unknownTag(0xffff, "Unknown tag", N_("Unknown tag"),
-                                    N_("Unknown tag"),
-                                    ifdIdNotSet, sectionIdNotSet, asciiString, -1, printValue);
-
     std::ostream& printValue(std::ostream& os, const Value& value, const ExifData*)
     {
         return os << value;
@@ -2923,361 +2712,3 @@ namespace Exiv2 {
     }
 
 }}                                      // namespace Internal, Exiv2
-
-namespace Exiv2 {
-
-    //! @cond IGNORE
-    GroupInfo::GroupName::GroupName(const std::string& groupName)
-    {
-        g_ = groupName;
-    }
-    //! @endcond
-
-    bool GroupInfo::operator==(int ifdId) const
-    {
-        return ifdId_ == ifdId;
-    }
-
-    bool GroupInfo::operator==(const GroupName& groupName) const
-    {
-        return 0 == strcmp(groupName.g_.c_str(), groupName_);
-    }
-
-    TagInfo::TagInfo(
-        uint16_t tag,
-        const char* name,
-        const char* title,
-        const char* desc,
-        int ifdId,
-        int sectionId,
-        TypeId typeId,
-        int16_t count,
-        PrintFct printFct
-    )
-        : tag_(tag), name_(name), title_(title), desc_(desc), ifdId_(ifdId),
-          sectionId_(sectionId), typeId_(typeId), count_(count), printFct_(printFct)
-    {
-    }
-
-    const GroupInfo* ExifTags::groupList()
-    {
-        return groupInfo + 1; // +1 to skip the first (ifdIdNotSet) entry
-    }
-
-    const TagInfo* ExifTags::tagList(const std::string& groupName)
-    {
-        const GroupInfo* ii = find(groupInfo, GroupInfo::GroupName(groupName));
-        if (ii == 0 || ii->tagList_ == 0) return 0;
-        return ii->tagList_();
-    } // ExifTags::tagList
-
-    const char* ExifTags::sectionName(const ExifKey& key)
-    {
-        const TagInfo* ti = tagInfo(key.tag(), static_cast<Internal::IfdId>(key.ifdId()));
-        if (ti == 0) return sectionInfo[unknownTag.sectionId_].name_;
-        return sectionInfo[ti->sectionId_].name_;
-    }
-
-    uint16_t ExifTags::defaultCount(const ExifKey& key)
-    {
-        const TagInfo* ti = tagInfo(key.tag(), static_cast<Internal::IfdId>(key.ifdId()));
-        if (ti == 0) return unknownTag.count_;
-        return ti->count_;
-    }
-
-    const char* ExifTags::ifdName(const std::string& groupName)
-    {
-        IfdId ifdId = Internal::groupId(groupName);
-        return Internal::ifdName(ifdId);
-    }
-
-    bool ExifTags::isMakerGroup(const std::string& groupName)
-    {
-        IfdId ifdId = Internal::groupId(groupName);
-        return Internal::isMakerIfd(ifdId);
-    }
-
-    bool ExifTags::isExifGroup(const std::string& groupName)
-    {
-        IfdId ifdId = Internal::groupId(groupName);
-        return Internal::isExifIfd(ifdId);
-    }
-
-    void ExifTags::taglist(std::ostream& os)
-    {
-        for (int i=0; ifdTagInfo[i].tag_ != 0xffff; ++i) {
-            os << ifdTagInfo[i] << "\n";
-        }
-        for (int i=0; exifTagInfo[i].tag_ != 0xffff; ++i) {
-            os << exifTagInfo[i] << "\n";
-        }
-        for (int i=0; iopTagInfo[i].tag_ != 0xffff; ++i) {
-            os << iopTagInfo[i] << "\n";
-        }
-        for (int i=0; gpsTagInfo[i].tag_ != 0xffff; ++i) {
-            os << gpsTagInfo[i] << "\n";
-        }
-    } // ExifTags::taglist
-
-    void ExifTags::taglist(std::ostream& os, const std::string& groupName)
-    {
-        IfdId ifdId = Internal::groupId(groupName);
-        Internal::taglist(os, ifdId);
-    }
-
-    //! %Internal Pimpl structure with private members and data of class ExifKey.
-    struct ExifKey::Impl {
-        //! @name Creators
-        //@{
-        Impl();                         //!< Default constructor
-        //@}
-
-        //! @name Manipulators
-        //@{
-        /*!
-          @brief Set the key corresponding to the \em tag, \em ifdId and \em tagInfo.
-                 The key is of the form '<b>Exif</b>.groupName.tagName'.
-         */
-        void makeKey(uint16_t tag, IfdId ifdId, const TagInfo* tagInfo);
-        /*!
-          @brief Parse and convert the key string into tag and IFD Id.
-                 Updates data members if the string can be decomposed,
-                 or throws \em Error .
-
-          @throw Error if the key cannot be decomposed.
-         */
-        void decomposeKey(const std::string& key);
-        //@}
-
-        //! @name Accessors
-        //@{
-        //! Return the name of the tag
-        std::string tagName() const;
-        //@}
-
-        // DATA
-        static const char* familyName_; //!< "Exif"
-
-        const TagInfo* tagInfo_;        //!< Tag info
-        uint16_t tag_;                  //!< Tag value
-        IfdId ifdId_;                   //!< The IFD associated with this tag
-        int idx_;                       //!< Unique id of the Exif key in the image
-        std::string groupName_;         //!< The group name
-        std::string key_;               //!< %Key
-    };
-
-    const char* ExifKey::Impl::familyName_ = "Exif";
-
-    ExifKey::Impl::Impl()
-        : tagInfo_(0), tag_(0), ifdId_(ifdIdNotSet), idx_(0)
-    {
-    }
-
-    std::string ExifKey::Impl::tagName() const
-    {
-        if (tagInfo_ != 0 && tagInfo_->tag_ != 0xffff) {
-            return tagInfo_->name_;
-        }
-        std::ostringstream os;
-        os << "0x" << std::setw(4) << std::setfill('0') << std::right
-           << std::hex << tag_;
-        return os.str();
-    }
-
-    void ExifKey::Impl::decomposeKey(const std::string& key)
-    {
-        // Get the family name, IFD name and tag name parts of the key
-        std::string::size_type pos1 = key.find('.');
-        if (pos1 == std::string::npos) throw Error(6, key);
-        std::string familyName = key.substr(0, pos1);
-        if (0 != strcmp(familyName.c_str(), familyName_)) {
-            throw Error(6, key);
-        }
-        std::string::size_type pos0 = pos1 + 1;
-        pos1 = key.find('.', pos0);
-        if (pos1 == std::string::npos) throw Error(6, key);
-        std::string groupName = key.substr(pos0, pos1 - pos0);
-        if (groupName.empty()) throw Error(6, key);
-        std::string tn = key.substr(pos1 + 1);
-        if (tn.empty()) throw Error(6, key);
-
-        // Find IfdId
-        IfdId ifdId = groupId(groupName);
-        if (ifdId == ifdIdNotSet) throw Error(6, key);
-        if (!Internal::isExifIfd(ifdId) && !Internal::isMakerIfd(ifdId)) {
-            throw Error(6, key);
-        }
-        // Convert tag
-        uint16_t tag = tagNumber(tn, ifdId);
-        // Get tag info
-        tagInfo_ = tagInfo(tag, ifdId);
-        if (tagInfo_ == 0) throw Error(6, key);
-
-        tag_ = tag;
-        ifdId_ = ifdId;
-        groupName_ = groupName;
-        // tagName() translates hex tag name (0xabcd) to a real tag name if there is one
-        key_ = familyName + "." + groupName + "." + tagName();
-    }
-
-    void ExifKey::Impl::makeKey(uint16_t tag, IfdId ifdId, const TagInfo* tagInfo)
-    {
-        assert(tagInfo != 0);
-
-        tagInfo_ = tagInfo;
-        tag_ = tag;
-        ifdId_ = ifdId;
-        key_ = std::string(familyName_) + "." + groupName_ + "." + tagName();
-    }
-
-    ExifKey::ExifKey(uint16_t tag, const std::string& groupName)
-        : p_(new Impl)
-    {
-        IfdId ifdId = groupId(groupName);
-        // Todo: Test if this condition can be removed
-        if (!Internal::isExifIfd(ifdId) && !Internal::isMakerIfd(ifdId)) {
-            throw Error(23, ifdId);
-        }
-        const TagInfo* ti = tagInfo(tag, ifdId);
-        if (ti == 0) {
-            throw Error(23, ifdId);
-        }
-        p_->groupName_ = groupName;
-        p_->makeKey(tag, ifdId, ti);
-    }
-
-    ExifKey::ExifKey(const TagInfo& ti)
-        : p_(new Impl)
-    {
-        IfdId ifdId = static_cast<IfdId>(ti.ifdId_);
-        if (!Internal::isExifIfd(ifdId) && !Internal::isMakerIfd(ifdId)) {
-            throw Error(23, ifdId);
-        }
-        p_->groupName_ = Exiv2::groupName(ifdId);
-        p_->makeKey(ti.tag_, ifdId, &ti);
-    }
-
-    ExifKey::ExifKey(const std::string& key)
-        : p_(new Impl)
-    {
-        p_->decomposeKey(key);
-    }
-
-    ExifKey::ExifKey(const ExifKey& rhs)
-        : Key(rhs), p_(new Impl(*rhs.p_))
-    {
-    }
-
-    ExifKey::~ExifKey()
-    {
-        delete p_;
-    }
-
-    ExifKey& ExifKey::operator=(const ExifKey& rhs)
-    {
-        if (this == &rhs) return *this;
-        Key::operator=(rhs);
-        *p_ = *rhs.p_;
-        return *this;
-    }
-
-    void ExifKey::setIdx(int idx)
-    {
-        p_->idx_ = idx;
-    }
-
-    std::string ExifKey::key() const
-    {
-        return p_->key_;
-    }
-
-    const char* ExifKey::familyName() const
-    {
-        return p_->familyName_;
-    }
-
-    std::string ExifKey::groupName() const
-    {
-        return p_->groupName_;
-    }
-
-    std::string ExifKey::tagName() const
-    {
-        return p_->tagName();
-    }
-
-    std::string ExifKey::tagLabel() const
-    {
-        if (p_->tagInfo_ == 0 || p_->tagInfo_->tag_ == 0xffff) return "";
-        return _(p_->tagInfo_->title_);
-    }
-
-    std::string ExifKey::tagDesc() const
-    {
-        if (p_->tagInfo_ == 0 || p_->tagInfo_->tag_ == 0xffff) return "";
-        return _(p_->tagInfo_->desc_);
-    }
-
-    TypeId ExifKey::defaultTypeId() const
-    {
-        if (p_->tagInfo_ == 0) return unknownTag.typeId_;
-        return p_->tagInfo_->typeId_;
-    }
-
-    uint16_t ExifKey::tag() const
-    {
-        return p_->tag_;
-    }
-
-    ExifKey::AutoPtr ExifKey::clone() const
-    {
-        return AutoPtr(clone_());
-    }
-
-    ExifKey* ExifKey::clone_() const
-    {
-        return new ExifKey(*this);
-    }
-
-    int ExifKey::ifdId() const
-    {
-        return p_->ifdId_;
-    }
-
-    int ExifKey::idx() const
-    {
-        return p_->idx_;
-    }
-
-    // *************************************************************************
-    // free functions
-
-    std::ostream& operator<<(std::ostream& os, const TagInfo& ti)
-    {
-        std::ios::fmtflags f( os.flags() );
-        ExifKey exifKey(ti);
-        os << exifKey.tagName() << ",\t"
-           << std::dec << exifKey.tag() << ",\t"
-           << "0x" << std::setw(4) << std::setfill('0')
-           << std::right << std::hex << exifKey.tag() << ",\t"
-           << exifKey.groupName() << ",\t"
-           << exifKey.key() << ",\t"
-           << TypeInfo::typeName(exifKey.defaultTypeId()) << ",\t"
-           << exifKey.tagDesc();
-
-        os.flags(f);
-        return os;
-    }
-
-}                                       // namespace Exiv2
-
-namespace {
-    std::ostream& printVersion(std::ostream& os, const std::string& str)
-    {
-        if (str.size() != 4) {
-            return os << "(" << str << ")";
-        }
-        if (str[0] != '0') os << str[0];
-        return os << str[1] << "." << str[2] << str[3];
-    }
-}
