+++ src/tiffcomposite_int.cpp
