+++ src/tiffvisitor_int.cpp
